feat: complete epic-01 doc/lint hygiene and bug fixes
- fix: compute_recent_observation_seed uses per-hydro max duration (F3-001)
- fix: partial-year inflow history validation excludes boundary years
- fix: CHANGELOG 18 duplicate blocks removed (1529→1025 lines)
- fix: CLAUDE.md test count 28→27, range D01-D11/D13-D16/D19-D30
- fix: 11 doc accuracy fixes across forward/backward/lp_builder/correlation
- fix: remove SDDP reference from cobre-solver (infrastructure genericity)
- fix: remove 5 stale dead_code suppressions, gate shift_lag_state with cfg(test)
- fix: replace noise_group_id_at silent u32::MAX fallback with debug_assert
- fix: add defensive debug_asserts for ring-buffer and negative-ID casts
- fix: promote MPI buffer alignment and spectral transform checks to assert
- fix: u32→u64 for angular pruning diagnostic counters

+++ crates/cobre-core/src/training_event.rs
@@ -227,10 +227,10 @@ pub enum TrainingEvent {
         /// Total number of cuts deactivated across all stages.
         cuts_deactivated: u32,
         /// Total number of angular clusters formed across all stages.
-        clusters_formed: u32,
+        clusters_formed: u64,
         /// Total number of within-cluster dominance checks performed across all
         /// stages.
-        dominance_checks: u32,
+        dominance_checks: u64,
         /// Number of stages processed (stages 1..num_stages-1; stage 0 is
         /// exempt).
         stages_processed: u32,
